refactor(demo): tighten editor app organization
