+++ apps/full-editor-demo/src/editor/shell/TopMenuBar.tsx
@@ -1,46 +1,20 @@
 import { useEffect, useState } from 'react';
 import { useTimelineState } from '@techsquidtv/canvas-timeline-react';
-import { ExportPanel } from '@/components/panels/ExportPanel';
 import { Button } from '@/components/ui/button';
-import { Separator } from '@/components/ui/separator';
 import { useSourceBin } from '@/components/source-bin/source-bin-context';
 import { demoProject } from '@/data/demo-project';
 import { useEditorProject, type ProjectAutosaveStatus } from '@/editor/project/project-context';
-import { formatRationalTime } from '@/lib/timeline-format';
 import { normalizeProjectTitle } from '@/project/project-metadata';
 import {
   defaultVideoResolutionPresetId,
   findVideoResolutionPreset,
-  formatVideoResolution,
   getVideoResolutionPresetId,
-  isVideoResolutionPresetId,
   type VideoResolutionPresetId,
-  videoResolutionPresets,
 } from '@/project/video-settings';
 import type { TimelineExportStatus } from '@/export/timeline-export-types';
-
-const aboutLinks = [
-  {
-    description: 'Project website and demos.',
-    href: 'https://canvastimeline.com',
-    label: 'Website',
-  },
-  {
-    description: 'Guides, API notes, and examples.',
-    href: 'https://canvastimeline.com/docs',
-    label: 'Documentation',
-  },
-  {
-    description: 'Full editor demo app source.',
-    href: 'https://github.com/TechSquidTV/Canvas-Timeline/tree/main/apps/full-editor-demo',
-    label: 'Demo source code',
-  },
-  {
-    description: 'Canvas Timeline monorepo.',
-    href: 'https://github.com/TechSquidTV/Canvas-Timeline',
-    label: 'GitHub repository',
-  },
-] as const;
+import { AboutMenu } from './top-menu/AboutMenu';
+import { ExportMenu } from './top-menu/ExportMenu';
+import { ProjectMenu } from './top-menu/ProjectMenu';
 
 type OpenMenu = 'about' | 'export' | 'project' | null;
 
@@ -122,101 +96,27 @@ export function TopMenuBar() {
             Project
           </Button>
           {openMenu === 'project' ? (
-            <div className="editor-menu-popover editor-project-menu">
-              <section className="editor-menu-section">
-                <h2 className="editor-menu-section-title">Current Project</h2>
-                <dl className="panel-readout">
-                  <div>
-                    <dt>Project</dt>
-                    <dd>{metadata.title}</dd>
-                  </div>
-                  <div>
-                    <dt>Resolution</dt>
-                    <dd>{formatVideoResolution(metadata)}</dd>
-                  </div>
-                  <div>
-                    <dt>Duration</dt>
-                    <dd>{formatRationalTime(state.duration)}</dd>
-                  </div>
-                  <div>
-                    <dt>Frame rate</dt>
-                    <dd>{metadata.frameRate} fps</dd>
-                  </div>
-                  <div>
-                    <dt>Sources</dt>
-                    <dd>{sources.length}</dd>
-                  </div>
-                </dl>
-              </section>
-              <Separator />
-              <section className="editor-menu-section">
-                <h2 className="editor-menu-section-title">New Project</h2>
-                {confirmingNewProject ? (
-                  <div className="editor-new-project-confirm">
-                    <label className="editor-field">
-                      <span>Project name</span>
-                      <input
-                        className="editor-input"
-                        onChange={(event) => setNewProjectTitleDraft(event.currentTarget.value)}
-                        value={newProjectTitleDraft}
-                      />
-                    </label>
-                    <label className="editor-field">
-                      <span>Resolution</span>
-                      <select
-                        className="editor-input"
-                        onChange={(event) => {
-                          if (isVideoResolutionPresetId(event.currentTarget.value)) {
-                            setNewProjectResolutionDraft(event.currentTarget.value);
-                          }
-                        }}
-                        value={newProjectResolutionDraft}
-                      >
-                        {videoResolutionPresets.map((preset) => (
-                          <option key={preset.id} value={preset.id}>
-                            {preset.label}
-                          </option>
-                        ))}
-                      </select>
-                    </label>
-                    <p>
-                      This deletes the current timeline, project settings, imported media, posters,
-                      and Source Bin entries.
-                    </p>
-                    {resetError === null ? null : <p className="editor-menu-error">{resetError}</p>}
-                    <div className="editor-menu-row">
-                      <Button
-                        disabled={resetting}
-                        onClick={() => {
-                          setConfirmingNewProject(false);
-                          setResetError(null);
-                          setNewProjectTitleDraft(demoProject.title);
-                          setNewProjectResolutionDraft(currentResolutionPresetId);
-                        }}
-                        variant="subtle"
-                      >
-                        Cancel
-                      </Button>
-                      <Button
-                        disabled={resetting || !storageAvailable}
-                        onClick={() => void confirmNewProject()}
-                        variant="subtle"
-                      >
-                        {resetting ? 'Deleting' : 'Delete and start new'}
-                      </Button>
-                    </div>
-                  </div>
-                ) : (
-                  <Button
-                    disabled={!storageAvailable}
-                    onClick={() => setConfirmingNewProject(true)}
-                    variant="subtle"
-                  >
-                    New Project
-                  </Button>
-                )}
-              </section>
-            </div>
+            <ProjectMenu
+              confirmingNewProject={confirmingNewProject}
+              duration={state.duration}
+              metadata={metadata}
+              newProjectResolutionDraft={newProjectResolutionDraft}
+              newProjectTitleDraft={newProjectTitleDraft}
+              onCancelNewProject={() => {
+                setConfirmingNewProject(false);
+                setResetError(null);
+                setNewProjectTitleDraft(demoProject.title);
+                setNewProjectResolutionDraft(currentResolutionPresetId);
+              }}
+              onConfirmNewProject={() => void confirmNewProject()}
+              onNewProjectResolutionDraftChange={setNewProjectResolutionDraft}
+              onNewProjectTitleDraftChange={setNewProjectTitleDraft}
+              onStartNewProject={() => setConfirmingNewProject(true)}
+              resetError={resetError}
+              resetting={resetting}
+              sourcesCount={sources.length}
+              storageAvailable={storageAvailable}
+            />
           ) : null}
         </div>
 
@@ -233,9 +133,7 @@ export function TopMenuBar() {
             Export
           </Button>
           {openMenu === 'export' ? (
-            <div className="editor-menu-popover editor-export-menu">
-              <ExportPanel status={exportStatus} onStatusChange={setExportStatus} />
-            </div>
+            <ExportMenu status={exportStatus} onStatusChange={setExportStatus} />
           ) : null}
         </div>
 
@@ -251,28 +149,7 @@ export function TopMenuBar() {
           >
             About
           </Button>
-          {openMenu === 'about' ? (
-            <div className="editor-menu-popover editor-about-menu">
-              <section className="editor-menu-section">
-                <h2 className="editor-menu-section-title">About Canvas Timeline</h2>
-                <div className="editor-menu-link-list">
-                  {aboutLinks.map((link) => (
-                    <a
-                      className="editor-menu-link"
-                      href={link.href}
-                      key={link.href}
-                      onClick={() => setOpenMenu(null)}
-                      rel="noreferrer"
-                      target="_blank"
-                    >
-                      <span>{link.label}</span>
-                      <small>{link.description}</small>
-                    </a>
-                  ))}
-                </div>
-              </section>
-            </div>
-          ) : null}
+          {openMenu === 'about' ? <AboutMenu onNavigate={() => setOpenMenu(null)} /> : null}
         </div>
       </nav>
