+++ apps/full-editor-demo/src/persistence/project/types.ts
@@ -1,6 +1,7 @@
 import type { Marker, TimelineClipGroup, Track } from '@techsquidtv/canvas-timeline-core';
 import type { RationalTime } from '@techsquidtv/canvas-timeline-utils';
-import type { EditorTrackKind } from '@/data/demo-project';
+import type { EditorTrackKind } from '../../data/demo-project';
+import type { ProjectMetadata } from '../../project/project-metadata';
 
 export interface PersistedTimelineState {
   clipGroups: TimelineClipGroup[];
@@ -17,23 +18,8 @@ export interface PersistedTimelineState {
   zoomScale: number;
 }
 
-export interface ProjectMetadata {
-  description: string;
-  frameRate: number;
-  height: number;
-  projectId: string;
-  title: string;
-  width: number;
-}
-
-export interface ProjectStorageSnapshot {
-  description: string;
-  frameRate: number;
-  height: number;
-  projectId: string;
+export interface ProjectStorageSnapshot extends ProjectMetadata {
   savedAt: string;
   timelineState: PersistedTimelineState;
-  title: string;
   version: 3;
-  width: number;
 }
