+++ apps/full-editor-demo/src/styles/timeline.css
@@ -1,13 +1,3 @@
-.full-editor-timeline-dock {
-  display: grid;
-  grid-template-rows: 3.25rem minmax(0, 1fr);
-}
-
-.full-editor-transport,
-.timeline-command-bar {
-  height: 3.25rem;
-}
-
 @layer components {
   .full-editor-timeline-dock {
     display: grid;
@@ -32,6 +22,7 @@
   .full-editor-transport {
     flex: 1 1 auto;
     min-width: 0;
+    height: 3.25rem;
     min-height: 3.25rem;
     overflow-x: auto;
     overflow-y: hidden;
@@ -46,6 +37,7 @@
     flex: 0 1 auto;
     max-width: min(44rem, 52%);
     min-width: 0;
+    height: 3.25rem;
     min-height: 3.25rem;
     gap: 0.375rem;
     overflow-x: auto;
